Reformat JSON response of categories value
The frontend application expects the categories value to be returned as
a dictionary of format: {"id", "type"}.

Previously, using an array changed the expected behavior of how
requests are sent, resulting in a number of off-by-one errors.
Instances where off-by-one errors were remedied by a manual fix have
been removed.